pyproject.toml
[project]
name = "agentcore-omnisearch-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.13.0",
]

[tool.uv.workspace]
members = [
    "apps/aws_knowledge_agent",
    "apps/omnisearch_agent",
    "apps/enduser_agent",
]
